style: fix type issue

+++ sekai/watch/custom_elements.py
@@ -52,7 +52,7 @@ class ScoreIndicator:
 
 @level_memory
 class LifeManager:
-    life: int
+    life: float
     initial_life: int
     max_life: int
     decrease_life: int
